Npc: Make maxhunger and maxthirst `long` instead of `short`
Fixes #1.

+++ src/Faction.h
@@ -503,10 +503,10 @@ class NPC
     short regtemp;
     short temp;
     short breathmax;
-    short maxhunger;
+    long maxhunger;
     float hunger;
     short nutrition;
-    short maxthirst;
+    long maxthirst;
     float thirst;
     short maxstamina;
     short stamina;
